backend/requirements.txt
Django==3.2.4
django_environ==0.4.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
gunicorn==20.1.0
django-phonenumber-field[phonenumberslite]==5.2.0
